fix issues

+++ pipeline_attend_and_excite.py
@@ -17,7 +17,6 @@ from diffusers.pipelines.pipeline_utils import DiffusionPipeline
 from diffusers.pipelines.stable_diffusion import StableDiffusionPipelineOutput
 from diffusers.pipelines.stable_diffusion.safety_checker import StableDiffusionSafetyChecker
 
-#from pipeline_sd import StableDiffusionPipeline
 from diffusers.pipelines.stable_diffusion import StableDiffusionPipeline
 
 from utils.gaussian_smoothing import GaussianSmoothing
@@ -208,7 +207,7 @@ class AttendAndExcitePipeline(StableDiffusionPipeline):
         for i in indices_to_alter:
             image = attention_for_text[:, :, i]
             if smooth_attentions:
-                smoothing = GaussianSmoothing(channels=1, kernel_size=kernel_size, sigma=sigma, dim=2).half().cuda()
+                smoothing = GaussianSmoothing(channels=1, kernel_size=kernel_size, sigma=sigma, dim=2).cuda()
                 input = F.pad(image.unsqueeze(0).unsqueeze(0), (1, 1, 1, 1), mode='reflect')
                 image = smoothing(input).squeeze(0).squeeze(0)
             max_indices_list.append(image.max())
